Phonepe/Datas/requirements.txt
folium==0.14.0
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
plotly==5.14.1
Requests==2.31.0
streamlit==1.22.0
streamlit_folium==0.12.0
streamlit_option_menu==0.3.6
